No Installation Argument Added and Others Updated
A new argument called -z has been added to allow the user to choose that the program not try to install the (python) packages it needs, which is helpful for super computers where the user doesn't have write access to where python is stored. This WILL require the user manually install all the other packages in requirements themselves though. 5 other arguments were updated as well so that they only need to be called to be toggled on/off. The README has been updated as well so that users are aware of this functionality.