README and File Names Update
Many of the file names have been updated for clarity on the end users part. Many parts of the final results have been actually added to the README finally too, which was overdue for an explanation. Explanations have also been updated, including for what networks, nodes, edges, and communities are.